remove training from onnx
