+++ test/integration/local/test_onnx.py
@@ -22,7 +22,7 @@ from test.integration import RESOURCE_PATH
 
 ONNX_PATH = os.path.join(RESOURCE_PATH, 'onnx')
 MODEL_PATH = os.path.join(ONNX_PATH, 'onnx_model')
-SCRIPT_PATH = os.path.join(MODEL_PATH, 'code', 'onnx_export_import.py')
+SCRIPT_PATH = os.path.join(MODEL_PATH, 'code', 'onnx_import.py')
 
 
 def test_onnx_import(docker_image, sagemaker_local_session, local_instance_type):
